+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for conan-flake
 
-## 0.11.0 (unreleased)
+## 0.11.0 (Aug 18, 2026)
 
 ### Documentation